SmartMatch/OCRContainer/requirements.txt
torch==2.10.0
torchvision==0.25.0
scikit-learn==1.7.2
numpy==2.2.6
scipy==1.15.3
Pillow==12.0.0
kraken==7.0
fastapi==0.133.1
uvicorn==0.38.0
python-multipart
tqdm
paddlepaddle==3.2.1
paddleocr[doc-parser]==3.4.0
paddlex[ocr]==3.4.3
transformers==4.57.6
sentencepiece==0.2.1
safetensors==0.7.0
filelock
PyYAML==6.0.2
natasha==1.6.0
